Moving the captcha check inside the account signup view, so the challenge is verified on the server.
