+++ hs_core/views/__init__.py
@@ -400,21 +400,6 @@ class CaptchaVerifyForm(forms.Form):
     response = forms.CharField()
 
 
-def verify_captcha(request):
-    f = CaptchaVerifyForm(request.POST)
-    if f.is_valid():
-        params = dict(f.cleaned_data)
-        params['privatekey'] = getattr(settings, 'RECAPTCHA_PRIVATE_KEY', '6LdNC_USAAAAADNdzytMK2-qmDCzJcgybFkw8Z5x')
-        params['remoteip'] = request.META['REMOTE_ADDR']
-        # return HttpResponse('true', content_type='text/plain')
-        resp = requests.post('http://www.google.com/recaptcha/api/verify', params=params)
-        lines = resp.text.split('\n')
-        if lines[0].startswith('false'):
-            raise ex.PermissionDenied('captcha failed')
-        else:
-            return HttpResponse('true', content_type='text/plain')
-
-
 def verify_account(request, *args, **kwargs):
     context = {
             'username' : request.GET['username'],
